+++ src/shapes/text.class.ts
@@ -1858,7 +1858,7 @@ export class Text<
     return this._fromObject(
       {
         ...object,
-        styles: stylesFromArray(object.styles, object.text),
+        styles: stylesFromArray(object.styles || {}, object.text),
       },
       {
         extraParam: 'text',