+++ test/unit/text.js
@@ -208,14 +208,15 @@
 
     fabric.Text.fromElement(elText, function(text) {
       assert.ok(text instanceof fabric.Text);
-      var expectedObject = fabric.util.object.extend(fabric.util.object.clone(REFERENCE_TEXT_OBJECT), {
+      var expectedObject = {
+        ...REFERENCE_TEXT_OBJECT,
         left: 0,
         top: -14.05,
         width: 8,
         height: 18.08,
         fontSize: 16,
         originX: 'left'
-      });
+      };
       assert.deepEqual(text.toObject(), expectedObject, 'parsed object is what expected');
     });
   });
@@ -249,7 +250,8 @@
 
       assert.ok(textWithAttrs instanceof fabric.Text);
 
-      var expectedObject = fabric.util.object.extend(fabric.util.object.clone(REFERENCE_TEXT_OBJECT), {
+      var expectedObject = {
+        ...REFERENCE_TEXT_OBJECT,
         /* left varies slightly due to node-canvas rendering */
         left:             fabric.util.toFixed(textWithAttrs.left + '', 2),
         top:              -88.03,
@@ -270,7 +272,7 @@
         fontWeight:       'bold',
         fontSize:         123,
         underline:        true,
-      });
+      };
       assert.deepEqual(textWithAttrs.toObject(), expectedObject);
     });
   });